requirements.txt
yfinance
pandas
numpy
scikit-learn
requests
lxml
beautifulsoup4
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
